@@ -11,8 +11,8 @@ namespace Countries.Model
     /// <summary>
     ///     A collection of coordinates that form a region (are neighbors of each other)
     /// </summary>
-    internal class Region
+    public class Region
     {
-        public List<Coordinate> Coordinates { get; } = new List<Coordinate>();
+        public List<Coordinate> Coordinates { get; } = new();
     }
 }
