@@ -11,7 +11,7 @@ namespace Countries.Model
     /// <summary>
     ///     X/Y position comparable by X/Y values
     /// </summary>
-    internal class Coordinate : IEquatable<Coordinate>
+    public sealed class Coordinate : IEquatable<Coordinate>
     {
         public Coordinate(int x, int y)
         {
@@ -22,7 +22,9 @@ namespace Countries.Model
         public int X { get; }
         public int Y { get; }
 
-        public bool Equals(Coordinate? other) => X.Equals(other?.X) && Y.Equals(other.Y);
+        public bool Equals(Coordinate? other) => other is { } coordinate && (X.Equals(coordinate.X) && Y.Equals(coordinate.Y));
+
+        public override bool Equals(object? obj) => obj is Coordinate coordinate && Equals(coordinate);
 
         public override int GetHashCode() => X + Y;
     }
