@@ -5,6 +5,7 @@
 // -----------------------------------------------
 
 using System;
+using System.Linq;
 using System.Text;
 using Countries.Support;
 
@@ -12,32 +13,48 @@ namespace Countries
 {
     internal static class Program
     {
+        private static readonly RegionCalculator RegionCalculator = new();
+        private static readonly Drawing          RegionDrawer     = new();
+
         private static void Main()
         {
             Console.OutputEncoding = Encoding.Unicode;
             Console.SetWindowSize(55, 50);
 
-            var run = new Runnable();
-
             int[][] specificData = TestData.GetSpecificLayout();
+            RegionDrawer.Draw("Predefined layout", specificData);
+            WriteResults(RegionCalculator.Calculate(specificData));
 
-            Console.WriteLine($"┌───────────────────┐");
-            Console.WriteLine($"│ Predefined layout │");
-            Console.WriteLine($"├{new string('─', 19)}┴─{new string('─', (specificData[0].Length * 2) - 19)}┐");
-            Console.WriteLine($"{Environment.NewLine}Total amount of regions: {run.Solution(specificData)}");
-            Console.WriteLine();
+            RandomRun(1);
 
-            int[][] randomData = TestData.GetRandomLayout();
+            var randomRuns = 1;
 
-            Console.WriteLine($"┌───────────────┐");
-            Console.WriteLine($"│ Random layout │");
-            Console.WriteLine($"├{new string('─', 15)}┴─{new string('─', (randomData[0].Length * 2) - 15)}┐");
-            Console.WriteLine($"{Environment.NewLine}Total amount of regions: {run.Solution(randomData)}");
-            Console.WriteLine();
-
-            Console.ReadKey();
+            while (Console.ReadKey().Key != ConsoleKey.Escape)
+            {
+                RandomRun(++randomRuns);
+            }
 
             Environment.Exit(0);
+        }
+
+        private static void RandomRun(int run)
+        {
+            int[][] randomData = TestData.GetRandomLayout();
+            RegionDrawer.Draw($"Random layout {run}", randomData);
+            WriteResults(RegionCalculator.Calculate(randomData));
+        }
+
+        private static void WriteResults(CalculationSummary summary)
+        {
+            if (summary.Results.Length == 1) Console.WriteLine("Only one country found!");
+
+            foreach ((int countryCode, int regions) in summary.Results)
+            {
+                Console.WriteLine($"Amount of regions of country ({Drawing.CountryFlags[countryCode]}): {regions}");
+            }
+
+            Console.WriteLine($"{Environment.NewLine}Total amount of regions: {summary.Results.Sum(result => result.AmountOfRegions)}");
+            Console.WriteLine();
         }
     }
 }
